Add support for multiple nonlinear variables in an equations set
This modifies the nonlinear mapping so that there may be multple nonlinear
variables, with a Jacobian matrix for each variable. The first nonlinear
variable passed to equations_mapping_residual_variable_types_set must have
the same number of dofs as the number of rows in the equations set, and
boundary conditions on this variable are used to decide if a row should be
kept in the solver Jacobian matrix.

The coupled finite elasticity and Darcy fluid pressure problem has been
modified so that the Darcy fluid pressure equation is now formulated as a
nonlinear equation to be able to take advantage of this change and use
sparse matrices.


git-svn-id: https://opencmiss.svn.sourceforge.net/svnroot/opencmiss/cm/trunk@2119 61e36070-c886-4827-b7c0-ed9e4dc59cc8